+++ src/UI.cpp
@@ -2,7 +2,7 @@
 **  Includes
 */
 
-#include "ui.hpp"
+#include "UI.hpp"
 
 /*
 **  Globals
